TG-63: Change file name Modal.message.js to Modal.js
